Curses::UI- perl
This test script is a script which takes an input from the command line (number of fields) and adds them to the currrent window. It also includes buttons to exit and change the page prev or next. I have also added a mechanism in order to set the tab order.